+++ packages/lexical/src/LexicalUpdates.ts
@@ -21,6 +21,7 @@ import type {
 import type {LexicalNode, SerializedLexicalNode} from './LexicalNode';
 import type {NodeParserState, ParsedNode} from './LexicalParsing';
 
+import getDOMSelection from 'shared/getDOMSelection';
 import invariant from 'shared/invariant';
 
 import {$isElementNode, $isTextNode} from '.';
@@ -39,13 +40,14 @@ import {
 import {initMutationObserver} from './LexicalMutations';
 import {$normalizeTextNode} from './LexicalNormalization';
 import {internalCreateNodeFromParse} from './LexicalParsing';
-import {updateEditorState} from './LexicalReconciler';
+import {reconcileRoot} from './LexicalReconciler';
 import {
   $isNodeSelection,
   $isRangeSelection,
   applySelectionTransforms,
   internalCreateSelection,
   internalCreateSelectionFromParse,
+  updateDOMSelection,
 } from './LexicalSelection';
 import {
   $getCompositionKey,
@@ -458,6 +460,10 @@ export function commitPendingUpdates(editor: LexicalEditor): void {
     return;
   }
 
+  // ======
+  // Reconcilation has started.
+  // ======
+
   const currentEditorState = editor._editorState;
   const currentSelection = currentEditorState._selection;
   const pendingSelection = pendingEditorState._selection;
@@ -468,53 +474,55 @@ export function commitPendingUpdates(editor: LexicalEditor): void {
   const previousReadOnlyMode = isReadOnlyMode;
   const previousActiveEditor = activeEditor;
   const previouslyUpdating = editor._updating;
+  const observer = editor._observer;
+  let mutatedNodes = null;
   activeEditor = editor;
   activeEditorState = pendingEditorState;
   isReadOnlyMode = false;
   // We don't want updates to sync block the reconcilation.
   editor._updating = true;
 
-  try {
-    if (!headless && rootElement !== null) {
-      const mutatedNodes = updateEditorState(
-        rootElement,
+  if (!headless && needsUpdate && observer !== null) {
+    try {
+      const dirtyType = editor._dirtyType;
+      const dirtyElements = editor._dirtyElements;
+      const dirtyLeaves = editor._dirtyLeaves;
+      observer.disconnect();
+
+      mutatedNodes = reconcileRoot(
         currentEditorState,
         pendingEditorState,
-        currentSelection,
-        pendingSelection,
-        needsUpdate,
         editor,
+        dirtyType,
+        dirtyElements,
+        dirtyLeaves,
       );
+    } catch (error) {
+      // Report errors
+      editor._onError(error);
 
-      if (mutatedNodes !== null) {
-        triggerMutationListeners(
-          editor,
-          currentEditorState,
-          pendingEditorState,
-          mutatedNodes,
-        );
+      // Reset editor and restore incoming editor state to the DOM
+      if (!isAttemptingToRecoverFromReconcilerError) {
+        resetEditor(editor, null, rootElement, pendingEditorState);
+        initMutationObserver(editor);
+        editor._dirtyType = FULL_RECONCILE;
+        isAttemptingToRecoverFromReconcilerError = true;
+        commitPendingUpdates(editor);
+        isAttemptingToRecoverFromReconcilerError = false;
       }
-    }
-  } catch (error) {
-    // Report errors
-    editor._onError(error);
 
-    // Reset editor and restore incoming editor state to the DOM
-    if (!isAttemptingToRecoverFromReconcilerError) {
-      resetEditor(editor, null, rootElement, pendingEditorState);
-      initMutationObserver(editor);
-      editor._dirtyType = FULL_RECONCILE;
-      isAttemptingToRecoverFromReconcilerError = true;
-      commitPendingUpdates(editor);
-      isAttemptingToRecoverFromReconcilerError = false;
+      return;
+    } finally {
+      observer.observe(rootElement, {
+        characterData: true,
+        childList: true,
+        subtree: true,
+      });
+      editor._updating = previouslyUpdating;
+      activeEditorState = previousActiveEditorState;
+      isReadOnlyMode = previousReadOnlyMode;
+      activeEditor = previousActiveEditor;
     }
-
-    return;
-  } finally {
-    editor._updating = previouslyUpdating;
-    activeEditorState = previousActiveEditorState;
-    isReadOnlyMode = previousReadOnlyMode;
-    activeEditor = previousActiveEditor;
   }
 
   pendingEditorState._readOnly = true;
@@ -522,6 +530,11 @@ export function commitPendingUpdates(editor: LexicalEditor): void {
   // @ts-ignore
   if (__DEV__) {
     handleDEVOnlyPendingUpdateGuarantees(pendingEditorState);
+    if ($isRangeSelection(pendingSelection)) {
+      Object.freeze(pendingSelection.anchor);
+      Object.freeze(pendingSelection.focus);
+    }
+    Object.freeze(pendingSelection);
   }
 
   const dirtyLeaves = editor._dirtyLeaves;
@@ -537,10 +550,45 @@ export function commitPendingUpdates(editor: LexicalEditor): void {
     editor._normalizedNodes = new Set();
     editor._updateTags = new Set();
   }
-
   $garbageCollectDetachedDecorators(editor, pendingEditorState);
-  const pendingDecorators = editor._pendingDecorators;
 
+  // ======
+  // Reconcilation has finished. Now update selection and trigger listeners.
+  // ======
+
+  const domSelection = headless ? null : getDOMSelection();
+
+  // Attempt to update the DOM selection, including focusing of the root element,
+  // and scroll into view if needed.
+  if (
+    !editor._readOnly &&
+    // domSelection will be null in headless
+    domSelection !== null &&
+    (needsUpdate || pendingSelection === null || pendingSelection.dirty)
+  ) {
+    activeEditorState = pendingEditorState;
+    try {
+      updateDOMSelection(
+        currentSelection,
+        pendingSelection,
+        editor,
+        domSelection,
+      );
+    } finally {
+      activeEditorState = previousActiveEditorState;
+    }
+  }
+
+  if (mutatedNodes !== null) {
+    triggerMutationListeners(
+      editor,
+      currentEditorState,
+      pendingEditorState,
+      mutatedNodes,
+    );
+  }
+
+  const pendingDecorators = editor._pendingDecorators;
   if (pendingDecorators !== null) {
     editor._decorators = pendingDecorators;
     editor._pendingDecorators = null;
